+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- RemoteKeyBuilder was missing the setRemoteRepository method.
 
 ## [0.11.7] - 2021-01-06
 ### Fixed